+++ lib/morandi/vips_image_processor.rb
@@ -7,6 +7,16 @@ require 'morandi/srgb_conversion'
 module Morandi
   # An alternative to ImageProcessor which is based on libvips for concurrent and less memory-intensive processing
   class VipsImageProcessor
+    # Colour filter related constants
+    RGB_LUMINANCE_EXTRACTION_FACTORS = [0.3086, 0.6094, 0.0820].freeze
+    SEPIA_MODIFIER = [25, 5, -25].freeze
+    BLUETONE_MODIFIER = [-10, 5, 25].freeze
+    COLOUR_FILTER_MODIFIERS = {
+      'sepia' => SEPIA_MODIFIER,
+      'bluetone' => BLUETONE_MODIFIER
+    }.freeze
+    SUPPORTED_FILTERS = COLOUR_FILTER_MODIFIERS.keys + ['greyscale']
+
     # Vips options are global, this method sets them for yielding, then restores to original
     def self.with_global_options(cache_max:, concurrency:)
       previous_cache_max = Vips.cache_max
@@ -50,6 +60,8 @@ module Morandi
         @scale = 1.0
       end
 
+      apply_filters!
+
       return unless @options['output.limit'] && @output_width && @output_height
 
       scale_factor = [@output_width, @output_height].max.to_f / [@img.width, @img.height].max
@@ -69,6 +81,32 @@ module Morandi
     end
 
     private
+
+    def apply_filters!
+      filter_name = @options['fx']
+      return unless SUPPORTED_FILTERS.include?(filter_name)
+
+      # The filter-related constants assume RGB colourspace, so it requires conversion
+      @img = @img.colourspace(:srgb) unless @img.interpretation == :srgb
+
+      # Convert to greyscale using weights
+      rgb_factors = RGB_LUMINANCE_EXTRACTION_FACTORS
+      recombination_matrix = [rgb_factors, rgb_factors, rgb_factors]
+      if @img.has_alpha?
+        # Add "0" multiplier for alpha to ignore it for luminance calculation
+        recombination_matrix = recombination_matrix.map { |channel_multipliers| channel_multipliers + [0] }
+        # Add fourth row in the matrix to preserve unchanged alpha channel
+        recombination_matrix << [0, 0, 0, 1]
+      end
+      @img = @img.recomb(recombination_matrix)
+
+      return unless COLOUR_FILTER_MODIFIERS[filter_name]
+
+      # Apply colour adjustment based on the modifiers setup
+      colour_filter_modifier = COLOUR_FILTER_MODIFIERS[filter_name]
+      colour_filter_modifier += [0] if @img.has_alpha?
+      @img = @img.linear(1.0, colour_filter_modifier)
+    end
 
     def not_equal_to_one(float)
       (float - 1.0).abs >= Float::EPSILON