+++ lib/morandi/profiled_pixbuf.rb
@@ -8,9 +8,7 @@ module Morandi
   # It attempts to load an image using jpegicc/littlecms to ensure that it is sRGB.
   # NOTE: pixbuf supports colour profiles, but it requires an explicit icc-profile option to embed it when saving file
   class ProfiledPixbuf < GdkPixbuf::Pixbuf
-    def initialize(path, local_options, max_size_px = nil)
-      @local_options = local_options
-
+    def initialize(path, _local_options, max_size_px = nil)
       path = srgb_path(path) || path
 
       if max_size_px
@@ -23,7 +21,7 @@ module Morandi
     private
 
     def srgb_path(original_path)
-      Morandi::SrgbConversion.perform(original_path, target_path: @local_options['path.icc'])
+      Morandi::SrgbConversion.perform(original_path)
     end
   end
 end