+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Vips colour filters
 - Vips crop
 - Vips rotate
+- Vips straighten
 
 ### Removed
 - [BREAKING] dropped support for a broken 'dominant' border colour