+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Vips crop
 - Vips rotate
 - Vips straighten
+- Vips gamma
 
 ### Removed
 - [BREAKING] dropped support for a broken 'dominant' border colour