+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.101.0] 14.01.2026
+### Removed
+- [BREAKING] support for custom srgb file path
+
 ## [0.100.0] 17.01.2024
 ### Added
 - Vips image processor (resizing)